Helpers for swapping; broken support for writing ELF
Changes probably look a little noisy because I partially converted
variable and function names from camel to snake case.

Using readelf, I see the right number of sections, but the section
header table is wrong somehow. Using hexdump, the strings that should be
there are not--so it seems like they aren't being written.